backend/requirements.txt
django==4.2
djangorestframework
drf_yasg
django-environ
psycopg2-binary
gunicorn